requirements.txt
tensorflow-gpu==2.4.1
tensorflow-hub==0.12.0
imageio ==2.9.0
imageio-ffmpeg==0.4.4
django==3.1.7
django-embed-video==1.4.0
django-jsonfield==1.4.1
pillow == 8.2.0
django-crispy-forms
dj-static
static-ranges